+++ src/bll/operation_log_test.go
@@ -10,8 +10,7 @@ import (
 )
 
 func TestOperationLog(t *testing.T) {
-	require := require.New(t)
-	tt := SetUpTestTools(require)
+	tt := SetUpTestTools(require.New(t))
 
 	uid := tpl.RandUID()
 	ctx := getUidContext(uid)
@@ -34,10 +33,11 @@ func TestOperationLog(t *testing.T) {
 			Percent: 2,
 		}
 		err := blls.OperationLog.Add(ctx, object, constant.OperationCreate, logContent)
-		require.Nil(err)
+		require.Nil(t, err)
 	})
 
 	t.Run("get operationLog", func(t *testing.T) {
+		require := require.New(t)
 		// 获取操作日志
 		page := &tpl.Pagination{}
 		page.Validate()
@@ -53,5 +53,45 @@ func TestOperationLog(t *testing.T) {
 		totalSize, err := daos.OperationLog.CountByObject(ctx, object)
 		require.Nil(err)
 		require.True(totalSize > 0)
+	})
+
+	t.Run("operationLog content", func(t *testing.T) {
+		require := require.New(t)
+		object := tpl.RandName()
+		logContent := &dto.OperationLogContent{
+			Users:   body.Users,
+			Groups:  body.Groups,
+			Desc:    tpl.RandName(),
+			Value:   body.Value,
+			Percent: 2,
+			Release: 111,
+		}
+		err := blls.OperationLog.Add(ctx, object, constant.OperationCreate, logContent)
+		require.Nil(err)
+
+		log, err := daos.OperationLog.FindOneByObject(ctx, object)
+		require.Nil(err)
+		require.Equal(logContent.Desc, log.Desc)
+
+		// 2
+		release := getRelease(log.Content)
+		require.Equal(int64(111), release)
+
+		// 3
+		item := &tpl.OperationLogListItem{}
+		parseLogContent(log.Content, item)
+
+		require.Equal(body.Users, item.Users)
+		require.Equal(body.Groups, item.Groups)
+		require.Equal(body.Value, item.Value)
+		require.Equal("userPercent", item.Kind)
+		require.Equal(2, item.Percent)
+	})
+
+	t.Run("FindAllByObject", func(t *testing.T) {
+		require := require.New(t)
+		logs, err := daos.OperationLog.FindAllByObject(nil)
+		require.Nil(err)
+		require.True(len(logs) > 0, len(logs))
 	})
 }