+++ src/schema/common.go
@@ -1,0 +1,6 @@
+package schema
+
+// CountResult ...
+type CountResult struct {
+	Count int `gorm:"column:count"`
+}