+++ src/service/services.go
@@ -1,0 +1,18 @@
+package service
+
+// Services ...
+type Services struct {
+	UrbsSetting UrbsSettingInterface
+	UserAuth    UserAuthInterface
+	GroupMember GroupMemberInterface
+}
+
+// NewServices ...
+func NewServices() *Services {
+	s := &Services{
+		GroupMember: &GroupMember{},
+		UrbsSetting: &UrbsSetting{},
+		UserAuth:    &UserAuth{},
+	}
+	return s
+}