+++ CHANGELOG.md
@@ -13,3 +13,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add test on github.
 - Add lock for sync group members.
 - Add ac UI.
+- support more query for user and group' settings API.